[🧹 REFACTOR] recognizeGesture 컴포넌트화
